Added separate header component
